# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dac": {
            "get": {
                "description": "Retrieves and processes DAC data from a local JSON file, grouping links by NPSN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migrate Digit 2025"
                ],
                "summary": "Get DAC data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dac.csv": {
            "get": {
                "description": "Retrieves DAC data from a local JSON file and returns it as a downloadable CSV file.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Migrate Digit 2025"
                ],
                "summary": "Get DAC data as CSV",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete": {
            "delete": {
                "description": "Deletes a file from the configured S3 bucket using folder and file query parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete file from S3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder inside the bucket",
                        "name": "folder",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File name to delete",
                        "name": "file",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get": {
            "get": {
                "description": "Downloads a file from the configured S3 bucket using folder and file query parameters.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Download file from S3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder inside the bucket",
                        "name": "folder",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File name to download",
                        "name": "file",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hisense": {
            "get": {
                "description": "Retrieves and processes Hisense data from a local JSON file, grouping links by NPSN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migrate Digit 2025"
                ],
                "summary": "Get Hisense data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hisense.csv": {
            "get": {
                "description": "Retrieves Hisense data from a local JSON file and returns it as a downloadable CSV file.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Migrate Digit 2025"
                ],
                "summary": "Get Hisense data as CSV",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "Lists all files in a specified folder within the configured S3 bucket and returns their download links.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List files in S3 folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder inside the bucket",
                        "name": "folder",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/zyrex": {
            "get": {
                "description": "Retrieves and processes Zyrex data from a local JSON file, grouping links by NPSN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migrate Digit 2025"
                ],
                "summary": "Get Zyrex data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}